pyproject.toml
[tool.poetry]
name = "bit-test-task"
version = "0.1.0"
description = ""
authors = ["Dmitriy Mikhalev <fakeforall@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
pydantic-extra-types = "^2.9.0"
simplejson = "^3.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
